+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v8.4.0-dev [2021-05-08]
 
+  - ### Store UI
+    - Disabled Last Updated sort by dropdown option (for now)
+
   - ### Patcher
     - Added HeaderBarButtons, lets you easily add buttons to the top right header, similar to ChannelTextAreaButtons