+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # GooseMod Changelog
 
-## v8.4.0-dev [2021-05-08]
+## v8.4.0 [2021-05-08]
 
   - ### Main Settings
     - Added option to force internal i18n GM lang